refactor: move historycode to SearchHistory.js
