+++ src/styles/productcard.css
@@ -23,7 +23,7 @@
   display: grid;
   justify-content: center;
   align-items: stretch;
-  background-color: #f4f4f4;
+  background-color: #ffffff;
 }
 
 .wide-container {
