+++ src/styles/productcard.scss
@@ -21,7 +21,7 @@
   display: grid;
   justify-content: center;
   align-items: stretch;
-  background-color: #f4f4f4;
+  background-color: #ffffff;
 }
 .wide-container {
   height: 100vh !important;
